@mlodato517 2022-01-25
> How did it figure that out?!

Haha you give me too much credit. I mean, technically, I didn't figure it out. I didn't actually examine the assembly for the case where this gets inlined into the `scan` iterator which, by Commit 4, I know to be optimized differently. But all I did was plug the two functions into `godbolt` (link in the PR description) and then read it for a really long time and googled every part that didn't make sense :sweat_smile:.

<details>
<summary>Slightly more detail?</summary>

Okay so maybe a step-by-step would be useful too :shrug: 
So here's the assembly output for the old code from godbolt:

```asm
example::test:
        mov     eax, 1   ; move 1 into the return register `eax`
        cmp     edi, 40  ; check if the input character is open paren
        je      .LBB0_3  ; if true, jump to `.LBBO_3` which returns (returning the 1 in `eax`)
        cmp     edi, 41  ; if it isn't, see if it's close paren
        jne     .LBB0_4  ; if it isn't, ump to `.LBBO_4` which aborts the program
        mov     eax, -1  ; if it is close paren, put -1 into `eax` and then return
.LBB0_3:
        ret
.LBB0_4: ; I'm going to ignore the details here - not sure why we bother with `rax` since we're aborting
         ; but this is basically "bail" by emitting `ud2` which is an "undefined instruction"
        push    rax
        call    qword ptr [rip + std::process::abort@GOTPCREL]
        ud2
```

and the assembly for the new code is:

```asm
example::test:
        xor     eax, eax  ; zero out the return register `eax`
        cmp     edi, 40   ; check if the input is open paren
        sete    al        ; Googled both `sete` and `al` to learn that `al` is the bottom
                          ; byte of the return register `eax` and that `sete` sets the register to 1
                          ; if the previous comparison returned true so this sets `eax` to 1` if the
                          ; input parameter is open paren
        add     eax, eax  ; double the `eax` register (it was 0/1 and is now 0/2)
        add     eax, -1   ; subtract one from `eax` (it's now -1/1)
        ret               ; return the value in `eax`
```

</details>